fix(gnovm): satisfy golangci-lint on the expansion guard
Two issues from the PR's lint job:
- predeclared: const max in expansionGas shadows the max builtin; renamed maxNodes.
- unconvert: int64(amount) in recordingGasMeter.ConsumeGas, where store.Gas is
  already int64.

+++ gnovm/pkg/gnolang/typecheck_bound_test.go
@@ -655,7 +655,7 @@ func newRecordingGasMeter(limit int64) *recordingGasMeter {
 }
 
 func (m *recordingGasMeter) ConsumeGas(amount store.Gas, descriptor string) {
-	m.charges = append(m.charges, int64(amount))
+	m.charges = append(m.charges, amount)
 	m.GasMeter.ConsumeGas(amount, descriptor)
 }
